Django_API/requirements.txt
Django==4.0.3
djangorestframework
django-cors-headers
djongo
dnspythons